+++ src/main/java/blackjack/view/OutputView.java
@@ -23,7 +23,6 @@ public class OutputView {
     private static final String PRINT_DEALER_NOT_RECEIVE_CARD = "\n딜러는 17이상이라 한장의 카드를 더 받지 못했습니다.\n";
     private static final String PRINT_FINAL_CARD_RESULT = "%s카드: %s - 결과: %d\n";
     private static final String PRINT_BLANK = " ";
-    private static final int COUNT_UNIT = 1;
 
     public static void printOpenCards(final Player dealer, final List<Gamer> gamers) {
         StringBuilder stringBuilder = new StringBuilder();
@@ -86,39 +85,20 @@ public class OutputView {
                 player.calculateResult());
     }
 
-    public static void printFinalResultBoard(final Map<Player, Result> gamerResultBoard) {
+    public static void printFinalResultBoard(final Map<Result, Integer> dealerResultBoard, final Map<Player, Result> gamerResultBoard) {
         System.out.println("\n## 최종 승패");
         System.out.printf(PRINT_DEFAULT_FORMAT_MESSAGE, Dealer.DEALER_NAME,
-                joinDealerString(gamerResultBoard));
+                joinDealerString(dealerResultBoard));
         gamerResultBoard.forEach((key, value) -> System.out.printf(PRINT_DEFAULT_FORMAT_MESSAGE,
                 key.getName(),
                 value.getResult()));
     }
 
-    private static String joinDealerString(final Map<Player, Result> gamerResultBoard) {
-        return calculateDealerResultBoard(gamerResultBoard).entrySet().stream()
+    private static String joinDealerString(final Map<Result, Integer> dealerResultBoard) {
+        return dealerResultBoard.entrySet().stream()
                 .map(board -> dealerResultToString(board.getKey(),
                         board.getValue()))
                 .collect(joining(PRINT_BLANK));
-    }
-
-    private static Map<Result, Integer> calculateDealerResultBoard(final Map<Player, Result> gamerResultBoard) {
-        Map<Result, Integer> enumMap = new EnumMap<>(Result.class);
-        for (Entry<Player, Result> gamerResultEntry : gamerResultBoard.entrySet()) {
-            Result dealerResult = convertToDealerResult(gamerResultEntry.getValue());
-            enumMap.put(dealerResult, enumMap.getOrDefault(dealerResult, 0) + COUNT_UNIT);
-        }
-        return enumMap;
-    }
-
-    private static Result convertToDealerResult(final Result result) {
-        if (result == Result.WIN) {
-            return Result.LOSE;
-        }
-        if (result == Result.LOSE) {
-            return Result.WIN;
-        }
-        return Result.DRAW;
     }
 
     private static String dealerResultToString(final Result result, final int value) {
